pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bedrock-usage-analyzer"
version = "0.1.0"
description = "Calculate token usage statistics and map quotas for Amazon Bedrock models"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Amazon Web Services"}
]
keywords = ["aws", "bedrock", "cloudwatch", "metrics", "tokens", "usage", "statistics"]

dependencies = [
    "boto3",
    "numpy",
    "jinja2",
    "pyyaml",
    "defusedcsv",
]

[project.urls]
Repository = "https://github.com/awslabs/bedrock-usage-analyzer"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bedrock_analyzer = ["templates/*.html"]